+++ src/firedays.h
@@ -17,30 +17,43 @@ enum fdEnumScreenMode {
 	FD_LEVELSCREEN = 4
 };
 
+typedef struct fdGameState {
+	int* screen_mode;
+	ngRect* screen_rect;
+	ngGrid* char_grid;
+	ngGrid* tile_grid;
+	ngGraphics* g;
+	ngMouse* mouse;
+	ngKey* key;
+} fdGameState;
+
 typedef struct fdTitleScreen {
 	ngRect header_rect;
 	ngRect button_rect;
 	ngRect start_rect;
 	ngRect quit_rect;
 } fdTitleScreen;
-void fd_title_screen_init (fdTitleScreen*);
-void fd_title_screen_draw (fdTitleScreen*, ngGraphics*);
+void fd_titlescreen_init (fdTitleScreen*, fdGameState*);
+void fd_titlescreen_event (fdTitleScreen*, fdGameState*);
+void fd_titlescreen_draw (fdTitleScreen*, fdGameState*);
 
 typedef struct fdFileScreen {
 	ngRect header_rect;
 	ngRect save_rect;
 	ngRect save1_rect;
 } fdFileScreen;
-void fd_file_screen_init (fdFileScreen*);
-void fd_file_screen_draw (fdFileScreen*, ngGraphics*);
+void fd_filescreen_init (fdFileScreen*, fdGameState*);
+void fd_filescreen_event (fdFileScreen*, fdGameState*);
+void fd_filescreen_draw (fdFileScreen*, fdGameState*);
 
 typedef struct fdWorldScreen {
 	ngRect header_rect;
 	ngRect level_rect;
 	ngRect level1_rect;
 } fdWorldScreen;
-void fd_world_screen_init (fdWorldScreen*);
-void fd_world_screen_draw (fdWorldScreen*, ngGraphics*);
+void fd_worldscreen_init (fdWorldScreen*, fdGameState*);
+void fd_worldscreen_event (fdWorldScreen*, fdGameState*);
+void fd_worldscreen_draw (fdWorldScreen*, fdGameState*);
 
 enum fdEnumHudMode { FD_HUDMENU = 1 };
 typedef struct fdHudMenu {
@@ -50,8 +63,9 @@ typedef struct fdHudMenu {
 	ngRect heat_rect;
 	ngRect fuel_rect;
 } fdHudMenu;
-void fd_hud_menu_init (fdHudMenu*);
-void fd_hud_menu_draw (fdHudMenu*, ngGraphics*);
+void fd_hudmenu_init (fdHudMenu*, fdGameState*);
+void fd_hudmenu_event (fdHudMenu*, fdGameState*);
+void fd_hudmenu_draw (fdHudMenu*, fdGameState*);
 
 enum fdEnumPauseMode { FD_PAUSEMENU = 1 };
 typedef struct fdPauseMenu {
@@ -61,16 +75,18 @@ typedef struct fdPauseMenu {
 	ngRect resume_rect;
 	ngRect quit_rect;
 } fdPauseMenu;
-void fd_pause_menu_init (fdPauseMenu*);
-void fd_pause_menu_draw (fdPauseMenu*, ngGraphics*);
+void fd_pausemenu_init (fdPauseMenu*, fdGameState*);
+void fd_pausemenu_event (fdPauseMenu*, fdGameState*);
+void fd_pausemenu_draw (fdPauseMenu*, fdGameState*);
 
 typedef struct fdLevelScreen {
 	ngRect header_rect;
 	fdHudMenu hud_menu;
 	fdPauseMenu pause_menu;
 } fdLevelScreen;
-void fd_level_screen_init (fdLevelScreen*);
-void fd_level_screen_draw (fdLevelScreen*, ngGraphics*);
+void fd_levelscreen_init (fdLevelScreen*, fdGameState*);
+void fd_levelscreen_event (fdLevelScreen*, fdGameState*);
+void fd_levelscreen_draw (fdLevelScreen*, fdGameState*);
 
 enum fdEnumDebugMode { FD_DEBUGMENU = 1 };
 typedef struct fdDebugMenu {
@@ -78,8 +94,9 @@ typedef struct fdDebugMenu {
 	ngRect version_rect;
 	ngRect fps_rect;
 } fdDebugMenu;
-void fd_debug_menu_init (fdDebugMenu*);
-void fd_debug_menu_draw (fdDebugMenu*, ngGraphics*);
+void fd_debugmenu_init (fdDebugMenu*, fdGameState*);
+void fd_debugmenu_event (fdDebugMenu*, fdGameState*);
+void fd_debugmenu_draw (fdDebugMenu*, fdGameState*);
 
 // Level